refactor(server): extract receiver / handle_check_seed_safety / build_init_response_data
The receiver closure (~130 lines) becomes `run_receiver(conn, world_engine,
image_gen, safety_checker, session, button_codes, system_info)` in
ws_session.py. handle_check_seed_safety follows it; build_init_response_data
becomes a small free function. `_update_gpu_metrics` and `build_binary_frame`
move onto Connection as `update_gpu_metrics` / `build_frame_envelope` methods
since both reference the cached-metrics fields.

server.py drops below 1100 lines for the first time. The big remaining piece
is the generator (~290 lines), extracted next.

Refs #78.
